requirements.txt
paho.mqtt.c
gattlib
libcurl4-openssl-dev
libreadline-gplv2-dev
